Frigging PCLATH now works. Buck it.

+++ PIC-Simu/PSInstruction.h
@@ -15,7 +15,7 @@
 @property (retain) NSString* instruction;
 @property (readwrite) int8_t registerAddress;
 @property (readwrite) int8_t bitAddress;
-@property (readwrite) int8_t literal;
+@property (readwrite) int16_t literal;
 @property (readwrite) BOOL storeInF;
 
 - (PSInstruction *)initWithBits:(int16_t)instructionBinary;
